fix: fixed delegate ownership feature + logout from navbar
