+++ app/components/navBar.tsx
@@ -4,15 +4,17 @@ import { signOut } from "firebase/auth";
 import { auth } from "../firebase/clientApp";
 import Image from "next/image";
 import Link from "next/link";
+import { useRouter } from "next/router";
 
 import { AuthContext } from "../context/AuthContext";
 
 const NavBar = () => {
   const appContext = useContext(AuthContext);
+  const router = useRouter();
 
   const signOutHandler = async () => {
     await signOut(auth);
-    document.location.reload();
+    router.push("/login");
   };
 
   return (
